// src/types/personal-pronoun.ts
export enum PersonalPronoun {
  /**
   *  - EN: I
   *  - HU: én
   *  - RU: я
   */
  FIRST_PERSON_SINGULAR = 'firstPersonSingular',
  /**
   * - EN: you
   * - HU: te
   * - RU: ты
   */
  SECOND_PERSON_SINGULAR = 'secondPersonSingular',
  /**
   * - EN: he/she/it
   * - HU: ő
   * - RU: он/она/оно
   */
  THIRD_PERSON_SINGULAR = 'thirdPersonSingular',
  /*
   * - EN: we
   * - HU: mi
   * - RU: мы
   */
  FIRST_PERSON_PLURAL = 'firstPersonPlural',
  /*
   * - EN: you
   * - HU: ti / önök (formal)
   * - RU: вы (formal)/ты (informal)
   */
  SECOND_PERSON_PLURAL = 'secondPersonPlural',
  /*
   * - EN: they
   * - HU: ők
   * - RU: они
   */
  THIRD_PERSON_PLURAL = 'thirdPersonPlural',
}